+++ src/configs/regexp.ts
@@ -16,7 +16,7 @@ export const regexpUnConfig: UnConfigFn<'regexp'> = (context) => {
   // 🟡 - in recommended (warns)
 
   configBuilder
-    ?.addConfig(['regexp', {includeDefaultFilesAndIgnores: true}])
+    ?.addConfig(['regexp', {includeDefaultFilesAndIgnores: true, doNotIgnoreHtml: true}])
     /* Category: Possible Errors */
     .addRule('no-contradiction-with-assertion', ERROR) // 🟢
     // "This rule is inspired by the `no-control-regex` rule. The positions of reports are improved over the core rule and suggestions are provided in some cases"