+++ type-decls.d.ts
@@ -39,3 +39,9 @@ declare module 'eslint-plugin-es-x' {
   declare const value: ESLint.ESLint.Plugin;
   export = value;
 }
+
+declare module 'eslint-plugin-html' {
+  import type ESLint from 'eslint';
+  declare const value: ESLint.ESLint.Plugin;
+  export = value;
+}